Introduce consistency check function everywhere
For each trait (PRF, Signature, Tweak Hash, Message Hash), we
introduce a function to check internal consistency of any given
parameters. This function is for testing only and it is assumed
to panic if something is wrong. This allows us to move certain
asserts from frequently called functions into tests.